refactor(compose): the capture overnight trio moves to capturejobs.go
jobs.go crossed the 500-LOC cap; the classify/enrich/digest job args and
worker adapters are one concept — the ADR-0063 overnight passes — so they
get their own file. Registration stays in NewJobRunner. No behaviour
change.
